platform/buffer_from_wl_shm: Handle clients immediately destroying the buffer.
We need to handle clients destroying the wl_buffer after commit. This is legal, as long
as the backing store is not invalidated.

Unfortunately, clients can create a buffer from the wl_shm_pool and then immediately
destroy the wl_shm_pool. This is also valid.

So in order to prevent the backing store from being invalidated we need to take a
reference on the wl_shm_pool backing our shm_buffer.

However!

If we have a reference on the wl_shm_pool, attempts to resize it will fail.
Real clients do this, so we can't universally take a reference to the pool.

Instead, we take a reference to the pool (and the buffer offset into it) when
the buffer is being destroyed. This way only clients that do the commit/delete
dance have their shm_pools frozen, and *those* clients will normally delete their
shm_pools immediately anyway.