platform/buffer_from_wl_shm: Handle out-of-bounds reads
It is not possible (in general) to guarantee up-front that access to a
client-provided SHM buffer will not generate a SIGBUS exception.

libwayland provides a tool to handle this - wl_shm_buffer_{begin,end}_access.
Unfortunately, this cannot be used after the wl_buffer has been destroyed,
as it crashes.

Also, this involves some annoying behind-our-backs manipulation of the
SIGBUS handler, for which we have a hack in run_mir to work around.

Since the libwayland functions don't actually work, and make our life
difficult, just don't use them! Instead, install our own SIGBUS handler
and handle this ourselves.